feat(settings): add user settings management and account deletion functionality
- Implemented phone number validation and formatting utility.
- Analyzed and documented state variable conversions to constants in settings logic.
- Created hooks for account deletion and email verification processes.
- Developed settings page logic to manage user profile, notifications, and Stripe integration.
- Added user profile management with fetching and updating capabilities.
- Integrated user notifications handling for email and SMS preferences.
- Enhanced user authentication with password change and reset functionalities.
- Updated layout component to support client-side rendering.
