+++ app/components/layout/ScreenHeaderWithBack.tsx
@@ -1,3 +1,4 @@
+"use client";
 import React from "react";
 import { ArrowLeft } from "lucide-react";
 import styles from "./ScreenHeaderWithBack.module.css";
